General update
Update for mpv's new VO option syntax and other things
