--- lua-settings/auto_options.conf
@@ -1,1 +1,0 @@
-verbose=no
