+++ lua-settings/osc.conf
@@ -4,6 +4,3 @@ valign=0.9
 deadzonesize=0.9
 hidetimeout=500
 minmousemove=1
-
-
-
