--- lua-settings/stats.conf
@@ -1,1 +1,0 @@
-debug=no
