+++ mpv.conf
@@ -24,6 +24,7 @@ screenshot-png-compression=8
 screenshot-template='~/Desktop/%F (%P) %n'
 
 hls-bitrate=max                         # use max quality for HLS streams
+ytdl-raw-options=format="rtmp-PREMIUM/h264_aac_ts_http_m3u8_http-3264/vp8_vorbis_webm_http_na_na-veryhigh/bestvideo+bestaudio/best"
 
 
 #########
@@ -31,7 +32,7 @@ hls-bitrate=max                         # use max quality for HLS streams
 #########
 
 cache=yes
-cache-default=50000                     # size in KB
+cache-default=800000                    # size in KB
 cache-backbuffer=25000                  # size in KB
 cache-initial=0                         # start playback when your cache is filled up with x kB
 cache-secs=10                           # how many seconds of audio/video to prefetch if the cache is active
@@ -45,13 +46,13 @@ osd-level=1                             # enable osd and display --osd-status-ms
 osd-duration=2500                       # hide the osd after x ms
 osd-status-msg='${time-pos} / ${duration}${?percent-pos:　(${percent-pos}%)}${?vo-drop-frame-count:${!vo-drop-frame-count==0:　Dropped: ${vo-drop-frame-count}}}\n${?chapter:Chapter: ${chapter}}'
 
-osd-font='Hiragino Maru Gothic Pro'
+osd-font='Source Sans Pro'
 osd-font-size=32
 osd-color='#CCFFFFFF'                   # ARGB format
-osd-border-color='#99000000'            # ARGB format
+osd-border-color='#DD322640'            # ARGB format
 #osd-shadow-offset=1                    # pixel width for osd text and progress bar
 osd-bar-align-y=0                       # progress bar y alignment (-1 top, 0 centered, 1 bottom)
-osd-border-size=1                       # size for osd text and progress bar
+osd-border-size=2                       # size for osd text and progress bar
 osd-bar-h=2                             # height of osd bar as a fractional percentage of your screen height
 osd-bar-w=60                            # width of " " "
 
@@ -61,6 +62,7 @@ osd-bar-w=60                            # width of " " "
 #############
 
 sub-use-margins
+ass-force-margins
 
 demuxer-mkv-subtitle-preroll            # try to correctly show embedded subs when seeking
 sub-auto=fuzzy                          # external subs don't have to match the file name exactly to autoload
@@ -94,32 +96,49 @@ alang=ja,jp,jpn,en,eng,de,deu,ger       # automatically select these audio track
 
 audio-file-auto=fuzzy                   # external audio doesn't has to match the file name exactly to autoload
 audio-pitch-correction=yes              # automatically insert scaletempo when playing with higher speed
-softvol=yes
-softvol-max=200                         # maximum volume in %, everything above 100 results in amplification
+volume-max=200                          # maximum volume in %, everything above 100 results in amplification
 volume=100                              # default volume, 100 = unchanged
 
 
 ################
-# Video Output # active VOs (and some other options) are set externally by a script
+# Video Output #
 ################
 
-vo=opengl-hq
-
+# Active VOs (and some other options) are set externally by a script
+# The profiles here are kind of a fallback
 
 [high-quality]
 profile-desc="High quality VO settings"
-vo=opengl-hq:scale=ewa_lanczossharp:cscale=ewa_lanczossoft:dscale=mitchell:tscale=triangle:scale-antiring=0.8:cscale-antiring=0.9:dither-depth=auto:target-prim=bt.709:correct-downscaling=yes
+vo=opengl
+profile=opengl-hq
+scale=ewa_lanczossharp
+cscale=ewa_lanczossoft
+dscale=mitchell
+tscale=linear
+scale-antiring=0.8
+cscale-antiring=0.9
 
 [mid-quality]
 profile-desc="Medium quality VO settings"
-vo=opengl-hq:scale=spline36:cscale=spline36:dscale=mitchell:tscale=triangle:scale-antiring=0.8:cscale-antiring=0.9:dither-depth=auto:target-prim=bt.709:correct-downscaling=yes
+vo=opengl
+profile=opengl-hq
+scale=spline36
+cscale=spline36
+dscale=mitchell
+tscale=linear
+scale-antiring=0.8
+cscale-antiring=0.9
 
 [low-quality]
 profile-desc="Low energy VO settings"
-vo=opengl:scale=spline36:dscale=mitchell:tscale=triangle:dither-depth=auto:target-prim=bt.709:correct-downscaling=yes
+vo=opengl
+scale=spline36
+dscale=mitchell
+tscale=linear
 hwdec=auto
 
 [default]
+vo=opengl
 
 
 ###################################
@@ -127,16 +146,16 @@ hwdec=auto
 ###################################
 
 [protocol.https]
-cache=yes
-cache-default=500000                    # size in KB
-cache-backbuffer=250000                 # size in KB
+#cache=yes
+#cache-default=500000                    # size in KB
+#cache-backbuffer=250000                 # size in KB
 cache-secs=100                          # how many seconds of audio/video to prefetch
 user-agent='Mozilla/5.0 (Macintosh; Intel Mac OS X 10.11; rv:47.0) Gecko/20100101 Firefox/47.0'
 
 [protocol.http]
-cache=yes
-cache-default=500000                    # size in KB
-cache-backbuffer=250000                 # size in KB
+#cache=yes
+#cache-default=500000                    # size in KB
+#cache-backbuffer=250000                 # size in KB
 cache-secs=100                          # how many seconds of audio/video to prefetch
 user-agent='Mozilla/5.0 (Macintosh; Intel Mac OS X 10.11; rv:47.0) Gecko/20100101 Firefox/47.0'
