+++ scripts/auto-options.lua
@@ -21,15 +21,14 @@ if mp.get_property_bool("option-info/vo/set-from-commandline") then
     return
 end
 
-local f = require 'auto-options-functions'
 local opts = require 'mp.options'
+local utils = require 'mp.utils'
 
 local o = {
     hq = "desktop",
     mq = "laptop",
-    lq = "laptop (low-power)",
-    highres_threshold = "1920:1200",
-    force_low_res = false,
+    lq = "low-energy",
+    highres_desktop_threshold = "1920:1200",
     verbose = false,
     duration = 5,
     duration_err_mult = 2,
@@ -37,88 +36,150 @@ local o = {
 opts.read_options(o)
 
 
--- Specify a VO for each level
+-- Specify mpv options for each level
 
-vo = {
-    [o.hq] = "opengl-hq",
-    [o.mq] = "opengl-hq",
-    [o.lq] = "opengl",
-}
-
-
--- Specify VO sub-options for different levels
-
-vo_opts = {
+local options = {
     [o.hq] = {
+        ["vo"] = "opengl",
         ["scale"]  = "ewa_lanczossharp",
         ["cscale"] = "ewa_lanczossoft",
         ["dscale"] = "mitchell",
-        ["tscale"] = "triangle",
+        ["tscale"] = "linear",
+        ["dither-depth"]        = "auto",
+        ["correct-downscaling"] = "yes",
+        ["sigmoid-upscaling"]   = "yes",
+        ["deband"]              = "yes",
         ["scale-antiring"]  = "0.8",
         ["cscale-antiring"] = "0.9",
 
-        ["dither-depth"]        = "auto",
-        ["target-prim"]         = "bt.709",
-        ["correct-downscaling"] = "yes",
+        ["hwdec"] = "no",
+        ["vd-lavc-threads"] = "16",
     },
 
     [o.mq] = {
+        ["vo"] = "opengl",
         ["scale"]  = "spline36",
-        ["cscale"] = "spline36",
         ["dscale"] = "mitchell",
-        ["tscale"] = "triangle",
+        ["tscale"] = "linear",
+        ["dither-depth"]        = "auto",
+        ["correct-downscaling"] = "yes",
+        ["sigmoid-upscaling"]   = "yes",
+        ["deband"]              = "yes",
         ["scale-antiring"]  = "0.8",
         ["cscale-antiring"] = "0.9",
 
-        ["dither-depth"]        = "auto",
-        ["target-prim"]         = "bt.709",
-        ["correct-downscaling"] = "yes",
+        ["hwdec"] = "no",
     },
 
     [o.lq] = {
+        ["vo"] = "opengl",
         ["scale"]  = "spline36",
         ["dscale"] = "mitchell",
-        ["tscale"] = "triangle",
+        ["tscale"] = "linear",
 
-        ["dither-depth"]        = "auto",
-        ["target-prim"]         = "bt.709",
-        ["correct-downscaling"] = "yes",
+        ["hwdec"] = "no",
     },
 }
 
 
--- Specify general mpv options for different levels
+-- Select the options level appropriate for this computer
+function determine_level(o, options)
+    -- Default level
+    local level = o.lq
 
-options = {
-    [o.hq] = {
-        ["options/vo"] = function () return vo_property_string(o.hq, vo, vo_opts) end,
-        ["options/hwdec"] = "no",
-        ["options/vd-lavc-threads"] = "16",
-    },
+    -- Overwrite level from command line with --script-opts=ao-level=<level>
+    local overwrite = mp.get_opt("ao-level")
+    if overwrite then
+        if not options[overwrite] then
+            print("Forced level does not exist: " .. overwrite)
+            return level
+        end
+        return overwrite
+    end
 
-    [o.mq] = {
-        ["options/vo"] = function () return vo_property_string(o.mq, vo, vo_opts) end,
-        ["options/hwdec"] = "no",
-    },
+    -- Call an external bash function determining whether this is a desktop or laptop
+    local loc = exec({"bash", "-c", 'source "$HOME"/local/shell/location-detection && is-desktop'})
+    if loc.error then
+        loc.status = 255
+    end
 
-    [o.lq] = {
-        ["options/vo"] = function () return vo_property_string(o.lq, vo, vo_opts) end,
-        ["options/hwdec"] = "auto",
-    },
-}
+    -- Desktop -> hq
+    if loc.status == 0 then
+        level = o.hq
+    -- Laptop  -> mq/lq
+    elseif loc.status == 1 then
+        level = o.mq
+        -- Go down to lq when we are on battery
+        bat = exec({"/usr/bin/pmset", "-g", "ac"})
+        if bat.stdout == "No adapter attached.\n" then
+            level = o.lq
+        end
+    elseif o.verbose then
+        print("unable to determine location, using default level: " .. level)
+    end
+
+    return level
+end
+
+
+-- Determine if the currently used resolution is higher than o.highres_threshold
+function high_res_desktop(o)
+    sp_ret = exec({"/usr/local/bin/resolution", "compare", o.highres_desktop_threshold})
+    return not sp_ret.error and sp_ret.status > 2
+end
+
+
+function high_res_video(o)
+    print("TODO: high_res_video(o)")
+    return false
+end
+
+
+function exec(process)
+    p_ret = utils.subprocess({args = process})
+    if p_ret.error and p_ret.error == "init" then
+        print("ERROR executable not found: " .. process[1])
+    end
+    return p_ret
+end
+
+
+function set_ASS(b)
+    return mp.get_property_osd("osd-ass-cc/" .. (b and "0" or "1"))
+end
+
+
+function red_border(s)
+    return set_ASS(true) .. "{\\bord1}{\\3c&H3300FF&}{\\3a&H20&}" .. s .. "{\\r}" .. set_ASS(false)
+end
+
+
+function print_status(name, value, o)
+    if not value or not o.level then
+        return
+    end
+
+    if o.err_occ then
+        print("Error setting level: " .. o.level)
+        mp.osd_message(red_border("Error setting level: ") .. o.level, o.duration * o.duration_err_mult)
+    else
+        print("Active level: " .. o.level)
+        mp.osd_message(o.level
+             .. (high_res_desktop(o) and "\n↳ desktop: high res" or "")
+             .. (high_res_video(o) and "\n↳ video: high res" or ""), o.duration)
+    end
+    mp.unobserve_property(print_status)
+end
 
 
 -- Print status information to VO window and terminal
-
 mp.observe_property("vo-configured", "bool",
                     function (name, value) print_status(name, value, o) end)
 
 
--- Determined level and set the appropriate options
-
+-- Determined level and apply the appropriate options
 function main()
-    o.force_low_res = mp.get_opt("ao-flr")
-    o.level = determine_level(o, vo, vo_opts, options)
+    o.level = determine_level(o, options)
     o.err_occ = false
     for k, v in pairs(options[o.level]) do
         if type(v) == "function" then
